Adding new enemy. Refatoring the code, now ships receive one scriptable object with the general information, treated internally. Player and enemys ships inherit from an abstract class named Ship.

+++ PainfulTestGame/Assets/GameAssets/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 { 
-      void TakeDamage(int damageValue);
+      void TakeDamage(float damageValue);
 }
